+++ shared/components/card/brand/_brand.scss
@@ -5,7 +5,7 @@ $headshot-sizes: (
 	S: 60
 );
 
-.card__opinion-header__image {
+.card__brand__image {
 	position: relative;
 	float: right;
 	margin: 0 0 0 oGridGutter();
@@ -18,7 +18,7 @@ $headshot-sizes: (
 	}
 }
 
-.card__opinion-header__image--headshot {
+.card__brand__image--headshot {
 	@include oColorsFor(author, background);
 
 	@each $breakpoint, $size in $headshot-sizes {
@@ -28,8 +28,7 @@ $headshot-sizes: (
 	}
 }
 
-.card__opinion-header__quote {
-	@include nextIcon(single-quote, oColorsGetColorFor(author, text), 30, $icon-only: true);
+.card__brand__quote {
 	position: absolute;
 	top: 0;
 	left: 0;
@@ -42,20 +41,21 @@ $headshot-sizes: (
 		@include respondTo($breakpoint) {
 			$quote-width: $size / 2;
 			$quote-height: $quote-width * (4/3);
+			@include nextIcon(single-quote, oColorsGetColorFor(author, text), $quote-width, $icon-only: true);
 			background-size: to-px($quote-width) to-px($quote-height);
 		}
 	}
 }
 
-.card__opinion-header__quote--left {
+.card__brand__quote--left {
 	background-position: 20% 50%;
 }
 
-.card__opinion-header__quote--right {
+.card__brand__quote--right {
 	background-position: 80% 50%;
 }
 
-.card__opinion-header__title {
+.card__brand__title {
 	font-family: $o-typography-serif-display;
 	font-weight: oFontsWeight(regular);
 	font-size: 20px;
@@ -63,7 +63,7 @@ $headshot-sizes: (
 	margin: 0 0 5px;
 }
 
-.card__opinion-header__link {
+.card__brand__link {
 	@include oColorsFor(author, text);
 	border-bottom: 0;
 }
