+++ shared/models/article.js
@@ -30,19 +30,19 @@ export default (item, opts) => {
 		if (item.authors.length || brand) {
 			const author = item.authors[0];
 			if (author && (author.isBrand || !brand)) {
-				article.opinionHeader = {
+				article.brand = {
 					title: author.name,
 					url: author.url,
 					headshot: author.headshot
 				};
 			} else {
-				article.opinionHeader = {
+				article.brand = {
 					title: brand.name,
 					url: brand.url
 				};
 			}
 		} else {
-			article.opinionHeader = {
+			article.brand = {
 				title: 'Opinion',
 				url: '/stream/sectionsId/MTE2-U2VjdGlvbnM='
 			};
